Fix error in adress-determination in find_data().

+++ src/toy.h
@@ -13,7 +13,7 @@
 
 int initialise_ram(uint16_t *ram, int argc, char **argv ); 
 uint8_t get_opcode(uint16_t instruction);
-int find_data(uint16_t instruction);
+uint16_t find_data(uint16_t instruction);
 int get2compl(uint16_t value); //not good place for something!
 bool execute(uint8_t op_code, int data_addr,uint16_t *ram);
